save download mode
Download mode is stored in r0. Now save it into [0xf9801000].

Now we can't return to OnChipRom as fastboot1, since bl1.bin is executed
in arm64 mode, not arm32 mode. We need to integrate fastboot protocol &
usb in bl1.bin instead.

And we still need to check whether it's download mode. So save it into
XG2RAM0 first.

Signed-off-by: Haojian Zhuang <haojian.zhuang@gmail.com>